x11: Align -m argument interpretation with dmenu
With dmenu, monitor indices start at 0 and a value of -1 (the default)
is used to spawn dmenu on the current monitor. While bemenu strives to
be compatible with dmenu, bemenu monitor indices previously started at 1
and a value of 0 (the default) was used to spawn on the current monitor.

This commit aligns the behaviour of bemenu's x11 backend with dmenu. For
this purposes, the affected code in the x11 backend is synced with the
current dmenu implementation. While doing so the monitor type has also
been switched from a uint32_t to a int32_t.

+++ lib/renderers/wayland/wayland.c
@@ -276,16 +276,16 @@ recreate_windows(const struct bm_menu *menu, struct wayland *wayland)
 {
     destroy_windows(wayland);
 
-    uint32_t monitors = 0;
+    int32_t monitors = 0;
     struct output *output;
     wl_list_for_each(output, &wayland->outputs, link)
         monitors++;
 
-    uint32_t monitor = 0;
+    int32_t monitor = 0;
     wl_list_for_each(output, &wayland->outputs, link) {
 
         if (!menu->monitor_name) {
-            if (menu->monitor != (uint32_t)-1) {
+            if (menu->monitor != -1) {
                  if (menu->monitor < monitors && monitor != menu->monitor) {
                     ++monitor;
                     continue;
@@ -318,7 +318,7 @@ recreate_windows(const struct bm_menu *menu, struct wayland *wayland)
         window->max_height = output->height;
         window->render_pending = true;
         wl_list_insert(&wayland->windows, &window->link);
-        if (menu->monitor != (uint32_t)-1) break;
+        if (menu->monitor != -1) break;
     }
 
     set_overlap(menu, menu->overlap);
@@ -332,7 +332,7 @@ fail:
 }
 
 static void
-set_monitor(const struct bm_menu *menu, uint32_t monitor)
+set_monitor(const struct bm_menu *menu, int32_t monitor)
 {
     (void)monitor;
     struct wayland *wayland = menu->renderer->internal;
